Cambiar el tipo de dato del Id de la entidad Zona.

+++ src/main/java/com/example/proyectlchsa/repositories/ZonaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ZonaRepository extends JpaRepository<Zona, Integer> {
+public interface ZonaRepository extends JpaRepository<Zona, Long> {
 }
